pyproject.toml
[project]
name = "scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
    "supermemory>=3.29.0",
]
